Move backend generated code to api-lib
