+++ scripts/db-seed.ts
@@ -9,8 +9,8 @@ const iti = (itiriri as unknown as { default: typeof itiriri }).default;
 import { DateTime, Duration, Interval } from 'luxon';
 
 import '../api-lib/node-fetch-shim';
+import { ValueTypes } from '../api-lib/gql/__generated__/zeus';
 import * as mutations from '../api-lib/gql/mutations';
-import { ValueTypes } from '../src/lib/gql/__generated__/zeusAdmin';
 
 const defaults = {
   seed: 9,
