+++ scripts/generate-zeus.sh
@@ -5,17 +5,19 @@ set -o allexport
 source .env
 set +o allexport
 
-GEN_PATH=./src/lib/gql/__generated__
+ADMIN_PATH=./api-lib/gql/__generated__
+USER_PATH=./src/lib/gql/__generated__
 
 function generate() {
   # use the first argument as the path to move files to
-  local MV_PATH=$GEN_PATH/$1; shift
+  local GEN_PATH=$1; shift
+
+  test -d "$GEN_PATH" && rm -r "$GEN_PATH"
 
   # pass the rest of the arguments to zeus
   (set -x; zeus "$HASURA_GRAPHQL_ENDPOINT"/v1/graphql "$GEN_PATH" --ts "$@")
-  test -d "$GEN_PATH"/zeus
-  test -d "$MV_PATH" && rm -r "$MV_PATH"
-  mv "$GEN_PATH"/zeus "$MV_PATH"
+#  test -d "$GEN_PATH"/zeus
+#  mv "$GEN_PATH"/zeus "$GEN_PATH"
 
 
 
@@ -31,19 +33,20 @@ function generate() {
   # for other seds try removing -i '' -e
   # use `brew install gsed` on macos to get this
   if [[ "$PLATFORM" == "OSX" || "$PLATFORM" == "BSD" ]]; then
-    sed -i "" 's,bigint"]:any,bigint"]:number,g' "$MV_PATH"/index.ts
-    sed -i "" 's,bigint"]:unknown,bigint"]:number,g' "$MV_PATH"/index.ts
+    sed -i "" 's,bigint"]:any,bigint"]:number,g' "$GEN_PATH"/zeus/index.ts
+    sed -i "" 's,bigint"]:unknown,bigint"]:number,g' "$GEN_PATH"/zeus/index.ts
   elif [ "$PLATFORM" == "LINUX" ]; then
-    sed -i 's,bigint"]:any,bigint"]:number,g' "$MV_PATH"/index.ts
-    sed -i 's,bigint"]:unknown,bigint"]:number,g' "$MV_PATH"/index.ts
+    sed -i 's,bigint"]:any,bigint"]:number,g' "$GEN_PATH"/zeus/index.ts
+    sed -i 's,bigint"]:unknown,bigint"]:number,g' "$GEN_PATH"/zeus/index.ts
   else
     echo "unknown platform; exiting"
     exit 1
   fi
 }
 
-generate zeusAdmin -h x-hasura-admin-secret:$HASURA_GRAPHQL_ADMIN_SECRET
-generate zeusUser -h x-hasura-role:user -h "authorization:generate"
+generate $ADMIN_PATH -h x-hasura-admin-secret:$HASURA_GRAPHQL_ADMIN_SECRET
+generate $USER_PATH -h x-hasura-role:user -h "authorization:generate"
 
 # fix formatting of generated files
-node_modules/.bin/prettier --write $GEN_PATH
+node_modules/.bin/prettier --write $ADMIN_PATH
+node_modules/.bin/prettier --write $USER_PATH
